+++ components/Post.js
@@ -1,13 +1,9 @@
-import { graphql } from "react-apollo";
-import gql from "graphql-tag";
-import { withRouter } from "next/router";
 import Head from "next/head";
 import Link from "next/link";
 
 import Datetime from "./Datetime";
 import ErrorMessage from "./ErrorMessage";
 import PostNav from "./PostNav";
-import WithError from "./WithError";
 import md from "../lib/markdown";
 import { logger } from "../lib/logger";
 
@@ -54,33 +50,3 @@ const Post = props => {
 
   return <div />;
 };
-
-Post.getInitialProps = async function(props) {
-  console.log(props)
-  return {statusCode: 404}
-};
-
-export const getPost = gql`
-  query getPost($id: ID!) {
-    post(id: $id) {
-      id
-      title
-      content
-      datetime
-      next {
-        id
-      }
-      prev {
-        id
-      }
-    }
-  }
-`;
-
-export default WithError(graphql(getPost, {
-  options: props => ({
-    variables: {
-      id: props.id,
-    },
-  }),
-})(withRouter(Post)));
