+++ pages/post.js
@@ -1,10 +1,13 @@
+import { withRouter } from "next/router";
+import { graphql } from "react-apollo";
+import gql from "graphql-tag";
+
+import { checkLoggedIn } from "../lib/auth";
 import App from "../components/App";
 import Header from "../components/Header";
 import Footer from "../components/Footer";
 import Post from "../components/Post";
-import { withRouter } from "next/router";
-
-import { checkLoggedIn } from "../lib/auth";
+import WithError from "../components/WithError";
 
 const Page = withRouter(props => (
   <App>
@@ -16,10 +19,43 @@ const Page = withRouter(props => (
 
 Page.getInitialProps = async ctx => {
   const { loggedInUser } = await checkLoggedIn(ctx.apolloClient);
+  console.log(ctx.data);
+  
+  if (ctx.data.error) {
+    return {
+      statusCode: 404
+    }
+  }
 
   return {
     loggedInUser,
+    id: ctx.query.id,
+    data: ctx.data
   };
 };
 
-export default Page;
+
+export const getPost = gql`
+  query getPost($id: ID!) {
+    post(id: $id) {
+      id
+      title
+      content
+      datetime
+      next {
+        id
+      }
+      prev {
+        id
+      }
+    }
+  }
+`;
+
+export default WithError(graphql(getPost, {
+  options: props => ({
+    variables: {
+      id: props.id,
+    },
+  }),
+})(Page));
